requirements.txt
Deprecated==1.2.13
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
jsonschema==4.6.0
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
pyrsistent==0.18.1
PyYAML==6.0
sqlparse==0.4.2
wrapt==1.14.1
